added newtype pattern for NodeId

+++ src/dijkstra.rs
@@ -1,6 +1,6 @@
 use std::{cmp::Ordering, collections::BinaryHeap};
 
-use crate::{fabric_graph::FabricGraph, node::NodeId};
+use crate::graph::{fabric_graph::FabricGraph, node::NodeId};
 
 impl FabricGraph {
     #[must_use]
@@ -14,7 +14,7 @@ impl FabricGraph {
 
         let mut max_frontier = 0usize;
 
-        dist[start as usize] = 0.0;
+        dist[start] = 0.0;
         heap.push(State {
             cost: 0.0,
             position: start ,
@@ -27,7 +27,7 @@ impl FabricGraph {
             }
 
             // If popped outdated distance, skip
-            if cost > dist[position as usize] {
+            if cost > dist[position] {
                 continue;
             }
 
@@ -38,7 +38,7 @@ impl FabricGraph {
 
                 while let Some(idx) = current {
                     path_indices.push(idx);
-                    current = prev[idx as usize];
+                    current = prev[idx];
                 }
 
                 path_indices.reverse();
@@ -47,14 +47,14 @@ impl FabricGraph {
             }
 
             // Expand adjacency list
-            for edge in &self.map[position as usize] {
+            for edge in &self.map[position] {
                 let base_cost = edge.cost;
-                let next_cost = cost + self.costs[edge.node_id as usize].calc_costs(base_cost, criticallity);
+                let next_cost = cost + self.costs[edge.node_id].calc_costs(base_cost, criticallity);
                 let next_pos = edge.node_id;
 
-                if next_cost < dist[next_pos as usize] {
-                    dist[next_pos as usize] = next_cost;
-                    prev[next_pos as usize] = Some(position ); 
+                if next_cost < dist[next_pos] {
+                    dist[next_pos] = next_cost;
+                    prev[next_pos] = Some(position ); 
                     heap.push(State {
                         cost: next_cost,
                         position: next_pos,
@@ -73,25 +73,25 @@ impl FabricGraph {
         let mut dist: Vec<f32> = vec![f32::MAX; n];
         let mut heap = BinaryHeap::new();
 
-        dist[start as usize] = 0.0;
+        dist[start] = 0.0;
         heap.push(State {
             cost: 0.0,
             position: start,
         });
 
         while let Some(State { cost, position }) = heap.pop() {
-            if cost > dist[position as usize] {
+            if cost > dist[position] {
                 continue;
             }
 
-            for edge in &self.map_reversed[position as usize] {
+            for edge in &self.map_reversed[position ] {
                 let base_cost = edge.cost;
-                let next_cost = cost + self.costs[edge.node_id as usize].calc_costs(base_cost, criticallity);
+                let next_cost = cost + self.costs[edge.node_id ].calc_costs(base_cost, criticallity);
 
                 let next_pos = edge.node_id;
 
-                if next_cost < dist[next_pos as usize] {
-                    dist[next_pos as usize] = next_cost;
+                if next_cost < dist[next_pos ] {
+                    dist[next_pos ] = next_cost;
                     heap.push(State {
                         cost: next_cost,
                         position: next_pos,
